+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Dependency Analysis Plugin Changelog
 
 # Version 0.44.0 (unreleased)
+* [New] Can now filter/configure compileOnly advice, analogous to other kinds of advice.
+(Thank you to Stephane Nicolas @stephanenicolas for contributing this!)
 * [Fixed] Build cache issue with `ArtifactsAnalysisTask`. (Thanks to @Paladinium for filing the 
 issue and providing a reproducer!)